feat: fill detector, stake limit live positions, stop order validation
- Add fill detector to track limit/stop order fills during fights
- Fix stake limit validation to use live Pacifica positions as fallback
  when FightTrade records are incomplete (fightSymbols filter)
- Update stake-info endpoint with same live position exposure logic
- Fix pendingNotional to use stop_price fallback for stop market orders
- Add stop order direction validation (buy stop > market, sell stop < market)
- Improve "Invalid stop tick" error messages from Pacifica
- Add $11 minimum order size check for all order types
- Fix stop/stop-limit tick size rounding with roundToTickSize
- Add trigger price placeholder hints showing required direction
- Use ::text cast for PostgreSQL enum queries (CREATE_STOP compatibility)
